change model

+++ Project_emotion/account/models.py
@@ -10,7 +10,7 @@ class User(AbstractUser):
     alram_sound = models.FloatField(default=0)
     led_color = models.CharField(max_length=40,default='')
     led_power = models.IntegerField(default=0)
-    led_bright = models.IntegerField(default=5)
+    led_bright = models.FloatField(default=5)
     def __str__(self):
         return str(self.username)
